+++ meson.options
@@ -1,3 +1,4 @@
 option('VERSION', type: 'string', value: '', description: 'Project version')
 option('tests', type: 'feature', value: 'auto', description: 'Build unit tests (requires cmocka)')
 option('nbdkit', type: 'feature', value: 'auto', description: 'Build the nbdkit plugin (requires nbdkit)')
+option('zip-container', type: 'feature', value: 'enabled', description: 'Flash from and create zip archives (requires libzip)')
