+++ src/file.c
@@ -5,71 +5,38 @@
 #include <fcntl.h>
 #include <stdlib.h>
 #include <unistd.h>
-#include <zip.h>
 
 #include "oscompat.h"
 #include "qdl.h"
 #include "file.h"
 
-struct qdl_zip {
-	zip_t *zip;
-	unsigned int refcount;
-};
-
 int qdl_file_open(struct qdl_zip *qdl_zip, const char *filename, struct qdl_file *file)
 {
-	struct zip_stat st;
-	zip_int64_t idx;
-	zip_file_t *zf;
 	off_t len;
-	zip_t *zip;
 	int fd;
 
-	if (qdl_zip) {
-		zip = qdl_zip->zip;
+	if (qdl_zip)
+		return qdl_zip_file_open(qdl_zip, filename, file);
 
-		idx = zip_name_locate(zip, filename, 0);
-		if (idx < 0) {
-			ux_err("unable to locate \"%s\" in zip archive\n", filename);
-			return -1;
-		}
-
-		if (zip_stat_index(zip, idx, 0, &st) < 0) {
-			ux_err("unable to stat \"%s\" in zip archive\n", filename);
-			return -1;
-		}
-
-		zf = zip_fopen_index(zip, idx, 0);
-		if (!zf) {
-			ux_err("unable to open \"%s\" in zip archive\n", filename);
-			return -1;
-		}
-
-		file->type = QDL_FILE_TYPE_ZIP;
-		file->fd = -1;
-		file->size = st.size;
-		file->zip_file = zf;
-	} else {
-		fd = open(filename, O_RDONLY | O_BINARY);
-		if (fd < 0) {
-			ux_err("failed to open \"%s\" for reading\n", filename);
-			return -1;
-		}
-
-		len = lseek(fd, 0, SEEK_END);
-		if (len < 0) {
-			ux_err("failed to find end of \"%s\"\n", filename);
-			close(fd);
-			return -1;
-		}
-
-		lseek(fd, 0, SEEK_SET);
-
-		file->type = QDL_FILE_TYPE_POSIX;
-		file->fd = fd;
-		file->size = len;
-		file->zip_file = NULL;
+	fd = open(filename, O_RDONLY | O_BINARY);
+	if (fd < 0) {
+		ux_err("failed to open \"%s\" for reading\n", filename);
+		return -1;
 	}
+
+	len = lseek(fd, 0, SEEK_END);
+	if (len < 0) {
+		ux_err("failed to find end of \"%s\"\n", filename);
+		close(fd);
+		return -1;
+	}
+
+	lseek(fd, 0, SEEK_SET);
+
+	file->type = QDL_FILE_TYPE_POSIX;
+	file->fd = fd;
+	file->size = len;
+	file->zip_file = NULL;
 
 	return 0;
 }
@@ -97,7 +64,7 @@ void *qdl_file_load(struct qdl_file *file, size_t *len)
 		}
 		break;
 	case QDL_FILE_TYPE_ZIP:
-		n = zip_fread(file->zip_file, buf, file->size);
+		n = qdl_zip_file_read(file, buf, file->size);
 		if ((size_t)n != file->size) {
 			ux_err("failed to load zip file member\n");
 			goto err_free_buf;
@@ -124,8 +91,7 @@ void qdl_file_close(struct qdl_file *file)
 		file->fd = -1;
 		break;
 	case QDL_FILE_TYPE_ZIP:
-		zip_fclose(file->zip_file);
-		file->zip_file = NULL;
+		qdl_zip_file_close(file);
 		break;
 	};
 
@@ -145,7 +111,7 @@ ssize_t qdl_file_read(struct qdl_file *file, void *buf, size_t len)
 	case QDL_FILE_TYPE_POSIX:
 		return read(file->fd, buf, len);
 	case QDL_FILE_TYPE_ZIP:
-		return zip_fread(file->zip_file, buf, len);
+		return qdl_zip_file_read(file, buf, len);
 	};
 
 	return -1;
@@ -195,47 +161,4 @@ off_t qdl_file_seek(struct qdl_file *file, off_t offset, int whence)
 	};
 
 	return -1;
-}
-
-int qdl_zip_open(const char *filename, struct qdl_zip **__qdl_zip)
-{
-	struct qdl_zip *qdl_zip;
-	zip_t *zip;
-
-	zip = zip_open(filename, ZIP_RDONLY, NULL);
-	if (!zip) {
-		*__qdl_zip = NULL;
-		return 0;
-	}
-
-	qdl_zip = calloc(1, sizeof(*qdl_zip));
-	if (!qdl_zip) {
-		zip_close(zip);
-		return -1;
-	}
-
-	qdl_zip->zip = zip;
-	qdl_zip->refcount = 1;
-
-	*__qdl_zip = qdl_zip;
-
-	return 0;
-}
-
-struct qdl_zip *qdl_zip_get(struct qdl_zip *qdl_zip)
-{
-	if (qdl_zip)
-		qdl_zip->refcount++;
-
-	return qdl_zip;
-}
-
-void qdl_zip_put(struct qdl_zip *qdl_zip)
-{
-	if (qdl_zip) {
-		if (--qdl_zip->refcount == 0) {
-			zip_close(qdl_zip->zip);
-			free(qdl_zip);
-		}
-	}
 }
